test: separate tests and test utils
